Refactor to allow new views
