+++ capture/a_capture.js
@@ -5,19 +5,16 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 */
 
-import { sendMessageToMain } from './utils.js'
+import { sendMessageToMain } from '../utils/utils.js'
 
-const WORKER_PREFIX = '[VIDEO-CAP]'
+const WORKER_PREFIX = '[AUDIO-CAP]'
 
 let stopped = false
 let mainLoopInterval
 let isMainLoopInExecution = false
 
-let timeCheck
-let estFps = 0
-
 function mainLoop (frameReader) {
-  return new Promise(function (resolve, reject) {
+  return new Promise(function (resolve) {
     if (isMainLoopInExecution) {
       return resolve(false)
     }
@@ -37,31 +34,19 @@ function mainLoop (frameReader) {
           sendMessageToMain(WORKER_PREFIX, 'info', 'Stream is done')
           return frameReader.cancel('ended')
         } else {
-          return new Promise(function (resolve, reject) { return resolve(result) })
+          return new Promise(function (resolve) { return resolve(result) })
         }
       }).then(result => {
         if (result === 'ended') {
           isMainLoopInExecution = false
           return resolve(false)
         } else {
-          const vFrame = result.value
+          const aFrame = result.value
+          sendMessageToMain(WORKER_PREFIX, 'debug', 'Read frame format: ' + aFrame.format + ', ts: ' + aFrame.timestamp + ', dur: ' + aFrame.duration + ', fs: ' + aFrame.sampleRate + ', Frames: ' + aFrame.numberOfFrames + ', ch: ' + aFrame.numberOfChannels)
 
-          sendMessageToMain(WORKER_PREFIX, 'debug', 'Read frame format: ' + vFrame.format + ', ts: ' + vFrame.timestamp + ', dur: ' + vFrame.duration)
-
-          // Send frame to process
-          self.postMessage({ type: 'vframe', clkms: Date.now(), data: vFrame }, [vFrame])
-          // vFrame.close();
-
-          estFps++
-          if (timeCheck === undefined) {
-            timeCheck = Date.now()
-          }
-          const nowMs = Date.now()
-          if (nowMs >= timeCheck + 1000) {
-            sendMessageToMain(WORKER_PREFIX, 'debug', 'estimated fps last sec: ' + estFps)
-            estFps = 0
-            timeCheck = nowMs
-          }
+          // AudioData is NOT transferable: https://github.com/WebAudio/web-audio-api/issues/2390
+          self.postMessage({ type: 'aframe', clkms: Date.now(), data: aFrame.clone() })
+          aFrame.close()
 
           isMainLoopInExecution = false
           return resolve(true)
@@ -81,15 +66,15 @@ self.addEventListener('message', async function (e) {
       sendMessageToMain(WORKER_PREFIX, 'error', 'Loop already running')
       return
     }
-    const vFrameStream = e.data.vStream
-    const vFrameReader = vFrameStream.getReader()
+    const aFrameStream = e.data.aStream
+    const aFrameReader = aFrameStream.getReader()
 
     sendMessageToMain(WORKER_PREFIX, 'info', 'Received streams from main page, starting worker loop')
 
-    mainLoopInterval = setInterval(mainLoop, 1, vFrameReader)
+    mainLoopInterval = setInterval(mainLoop, 1, aFrameReader)
 
     return
   }
 
-  sendMessageToMain(WORKER_PREFIX, 'error', 'Invalid message received.')
+  sendMessageToMain(WORKER_PREFIX, 'error', 'Invalid message received')
 })
