+++ capture/v_capture.js
@@ -5,16 +5,19 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 */
 
-import { sendMessageToMain } from './utils.js'
+import { sendMessageToMain } from '../utils/utils.js'
 
-const WORKER_PREFIX = '[AUDIO-CAP]'
+const WORKER_PREFIX = '[VIDEO-CAP]'
 
 let stopped = false
 let mainLoopInterval
 let isMainLoopInExecution = false
 
+let timeCheck
+let estFps = 0
+
 function mainLoop (frameReader) {
-  return new Promise(function (resolve, reject) {
+  return new Promise(function (resolve) {
     if (isMainLoopInExecution) {
       return resolve(false)
     }
@@ -34,19 +37,31 @@ function mainLoop (frameReader) {
           sendMessageToMain(WORKER_PREFIX, 'info', 'Stream is done')
           return frameReader.cancel('ended')
         } else {
-          return new Promise(function (resolve, reject) { return resolve(result) })
+          return new Promise(function (resolve) { return resolve(result) })
         }
       }).then(result => {
         if (result === 'ended') {
           isMainLoopInExecution = false
           return resolve(false)
         } else {
-          const aFrame = result.value
-          sendMessageToMain(WORKER_PREFIX, 'debug', 'Read frame format: ' + aFrame.format + ', ts: ' + aFrame.timestamp + ', dur: ' + aFrame.duration + ', fs: ' + aFrame.sampleRate + ', Frames: ' + aFrame.numberOfFrames + ', ch: ' + aFrame.numberOfChannels)
+          const vFrame = result.value
 
-          // AudioData is NOT transferable: https://github.com/WebAudio/web-audio-api/issues/2390
-          self.postMessage({ type: 'aframe', clkms: Date.now(), data: aFrame.clone() })
-          aFrame.close()
+          sendMessageToMain(WORKER_PREFIX, 'debug', 'Read frame format: ' + vFrame.format + ', ts: ' + vFrame.timestamp + ', dur: ' + vFrame.duration)
+
+          // Send frame to process
+          self.postMessage({ type: 'vframe', clkms: Date.now(), data: vFrame }, [vFrame])
+          // vFrame.close();
+
+          estFps++
+          if (timeCheck === undefined) {
+            timeCheck = Date.now()
+          }
+          const nowMs = Date.now()
+          if (nowMs >= timeCheck + 1000) {
+            sendMessageToMain(WORKER_PREFIX, 'debug', 'estimated fps last sec: ' + estFps)
+            estFps = 0
+            timeCheck = nowMs
+          }
 
           isMainLoopInExecution = false
           return resolve(true)
@@ -66,15 +81,15 @@ self.addEventListener('message', async function (e) {
       sendMessageToMain(WORKER_PREFIX, 'error', 'Loop already running')
       return
     }
-    const aFrameStream = e.data.aStream
-    const aFrameReader = aFrameStream.getReader()
+    const vFrameStream = e.data.vStream
+    const vFrameReader = vFrameStream.getReader()
 
     sendMessageToMain(WORKER_PREFIX, 'info', 'Received streams from main page, starting worker loop')
 
-    mainLoopInterval = setInterval(mainLoop, 1, aFrameReader)
+    mainLoopInterval = setInterval(mainLoop, 1, vFrameReader)
 
     return
   }
 
-  sendMessageToMain(WORKER_PREFIX, 'error', 'Invalid message received')
+  sendMessageToMain(WORKER_PREFIX, 'error', 'Invalid message received.')
 })
