+++ src-encoder/simple.html
@@ -107,7 +107,7 @@ LICENSE file in the root directory of this source tree.
 
     function createWorkers() {
         // Create send worker
-        muxerSenderWorker = new Worker("./moq_sender.js", { type: "module" } );
+        muxerSenderWorker = new Worker(".../sender/moq_sender.js", { type: "module" } );
     }
 
     function initUI() {
