+++ src-player/simple.html
@@ -123,7 +123,7 @@ LICENSE file in the root directory of this source tree.
 
     function createWorkers() {
         // Create a worker to download chunk
-        muxerDownloaderWorker = new Worker("./moq_demuxer_downloader.js", {type: "module"});
+        muxerDownloaderWorker = new Worker("../receiver/moq_demuxer_downloader.js", {type: "module"});
     }
 
     function addReceivedMessage(str) {
