+++ utils/utils.js
@@ -5,6 +5,8 @@ This source code is licensed under the MIT license found in the
 LICENSE file in the root directory of this source tree.
 */
 
+'use strict'
+
 export class StateEnum {
   // Create new instances of the same class as static attributes
   static Created = new StateEnum('created')
@@ -18,10 +20,42 @@ export class StateEnum {
 }
 
 export function sendMessageToMain (prefix, type, data) {
-  if (type === 'debug' || type === 'info' || type === 'error' || type === 'warning') {
-    data = prefix + ' ' + data
+    if (type === 'debug' || type === 'info' || type === 'error' || type === 'warning') {
+      data = prefix + ' ' + data
+    }
+    self.postMessage({ type, data })
+}
+
+export function serializeMetadata (metadata) {
+  let ret
+  if (isMetadataValid(metadata)) {
+    const newData = {}
+    // Copy all enumerable own properties
+    newData.decoderConfig = Object.assign({}, metadata.decoderConfig)
+    // Description is buffer
+    if ('description' in metadata.decoderConfig) {
+      newData.decoderConfig.descriptionInBase64 = arrayBufferToBase64(metadata.decoderConfig.description)
+      delete newData.description
+    }
+    // Encode
+    const encoder = new TextEncoder()
+    ret = encoder.encode(JSON.stringify(newData))
   }
-  self.postMessage({ type, data })
+  return ret
+}
+
+export function isMetadataValid (metadata) {
+  return metadata !== undefined && 'decoderConfig' in metadata
+}
+
+function arrayBufferToBase64 (buffer) {
+  let binary = ''
+  const bytes = new Uint8Array(buffer)
+  const len = bytes.byteLength
+  for (let i = 0; i < len; i++) {
+    binary += String.fromCharCode(bytes[i])
+  }
+  return btoa(binary)
 }
 
 export function deSerializeMetadata (metadata) {
